+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased Changes]
 - [Parse command line arguments](https://github.com/amp-lang/ampc/issues/11)
+- [Disable returning types from functions entirely](https://github.com/amp-lang/ampc/issues/10)
 
 ### [`snapshot-05-06-2023-a`](https://github.com/amp-lang/ampc/releases/tag/snapshot-05-06-2023-a)
 - [Report diagnostics in `Type::from_ast`](https://github.com/amp-lang/ampc/issues/5) (report invalid types)